@@ -1,12 +1,9 @@
-﻿using System;
+﻿namespace Blazor.ComponentsSample.Shared;
 
-namespace Blazor.ComponentsSample.Shared
+public class Book
 {
-    public class Book
-    {
-        public Guid Id { get; set; }
-        public string Title { get; set; } = string.Empty;
-        public string? Publisher { get; set; }
-        public DateTime ReleaseDate { get; set; }
-    }
+    public Guid Id { get; set; }
+    public string Title { get; set; } = string.Empty;
+    public string? Publisher { get; set; }
+    public DateTime ReleaseDate { get; set; }
 }
