@@ -1,8 +1,10 @@
-@page "/counter"
+﻿@page "/counter"
+
+<PageTitle>Counter</PageTitle>
 
 <h1>Counter</h1>
 
-<p>Current count: @currentCount</p>
+<p role="status">Current count: @currentCount</p>
 
 <button class="btn btn-primary" @onclick="IncrementCount">Click me</button>
 
