@@ -1,4 +1,4 @@
-@page "/counterwithservice"
+﻿@page "/counterwithservice"
 @inject CounterService CounterService
 
 <h1>Counter</h1>
