@@ -7,6 +7,6 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using Blazor.ComponentsSample
+@using Blazor.ComponentsSample.Models
 @using Blazor.ComponentsSample.Shared
-@using Blazor.ComponentsSample.Models 
-@using Blazor.ComponentsSample.Client.Services 
+@using Blazor.ComponentsSample.Services
