@@ -1,4 +1,6 @@
-@page "/"
+﻿@page "/"
+
+<PageTitle>Index</PageTitle>
 
 <h1>Hello, world!</h1>
 
