@@ -1,17 +1,19 @@
-@inherits LayoutComponentBase
+﻿@inherits LayoutComponentBase
+
+<PageTitle>Blazor.ServerSample</PageTitle>
 
 <div class="page">
     <div class="sidebar">
         <NavMenu />
     </div>
 
-    <div class="main">
+    <main>
         <div class="top-row px-4">
             <a href="https://docs.microsoft.com/aspnet/" target="_blank">About</a>
         </div>
 
-        <div class="content px-4">
+        <article class="content px-4">
             @Body
-        </div>
-    </div>
+        </article>
+    </main>
 </div>
