@@ -1,17 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+namespace Blazor.WasmSample.Shared;
 
-namespace Blazor.WasmSample.Shared
+public class WeatherForecast
 {
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
+    public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+    public int TemperatureC { get; set; }
 
-        public string? Summary { get; set; }
+    public string? Summary { get; set; }
 
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-    }
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 }
